fix: PBH signature aggregator address
