+++ contracts/test/PBHSignatureAggregator.t.sol
@@ -242,45 +242,6 @@ contract PBHSignatureAggregatorTest is TestSetup {
         assertEq(decodedProofs[1].proof[7], proof.proof[7], "Proof should match");
     }
 
-    // PackedUserOperation { 
-    // sender: 0x4dfe55bc8eea517efdbb6b2d056135b350b79ca2, 
-    // nonce: 98268500254511791751890014262541266044804674709396859921428445945272626839552, 
-    // initCode: 0x, 
-    // callData: 0x7bb3742800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000, 
-    // accountGasLimits: 0x000000000000000000000000000fffd30000000000000000000000000000c350, 
-    // preVerificationGas: 60232, 
-    // gasFees: 0x0000000000000000000000003b9aca0000000000000000000000000053140b60, 
-    // paymasterAndData: 0x, 
-    // signature: 0x000000000000000000000000fd2e93e8fd5c92c4a4ba02047f18efdffda1fc7b8a155c6a9730434be0bbe6a9489098947008a2e2988f3ae11218874e7a773ed371b1138e07f8970ebcdd12502005276ad6d825269eb0b67a2e1589123ded27c8b8eabfa898ff7e878ad61071ad01000c07e80000000000000000000000000000000000000000000000000000002fdc7e1f7d7acf3c83bb6b9e25053ddef53a62ee0e6aa97e847c2cdd1d3b670626328b2cb62fe27833c589ab8a900c202fb0e019de76ab7e50b6d87098a3cebc2f79c04f0737cf4551bc209457599e0806b6e36ce9725bddf06a7aeb0fd5ed7127a02c48981c90ed1e47f9492d32a73c930eea823fb8ce7a23f73cff3381890c01c94797dfa4fbc9b1b226d65972c80acf1d10a471801c48aeccbd7b325b50b408b20977c6445febc6a1c8a849bdbc64b4497f08b141418164b31bc999b9869f01f138a142aa38e4454d217e8532290f20e211c5578639f0f7a8a381e02f667020bffc27977f2ca0a735ccf8d946b981443d1026a8b70be166b50005e0357789067cd8f68081a941e1068961830f36a73e24cbc2cd1dac1cb4a781475b5a95c3 
-    // }
-    function test_validateUserOperation() public {
-        Mock4337Module mockModule = Mock4337Module(address(0x0B306BF915C4d645ff596e518fAf3F9669b97016));
-        PackedUserOperation memory userOp = PackedUserOperation({
-            sender: address(0x4DFE55bC8eEa517efdbb6B2d056135b350b79ca2),
-            nonce: 98268500254511791751890014262541266044804674709396859921428445945272626839552,
-            initCode: new bytes(0),
-            callData: hex"7bb3742800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-            accountGasLimits: 0x000000000000000000000000000fffd30000000000000000000000000000c350,
-            preVerificationGas: 60232,
-            gasFees: 0x0000000000000000000000003b9aca0000000000000000000000000053140b60,
-            paymasterAndData: new bytes(0),
-            signature: hex"000000000000000000000000fd2e93e8fd5c92c4a4ba02047f18efdffda1fc7b8a155c6a9730434be0bbe6a9489098947008a2e2988f3ae11218874e7a773ed371b1138e07f8970ebcdd12502005276ad6d825269eb0b67a2e1589123ded27c8b8eabfa898ff7e878ad61071ad01000c07e80000000000000000000000000000000000000000000000000000002fdc7e1f7d7acf3c83bb6b9e25053ddef53a62ee0e6aa97e847c2cdd1d3b670626328b2cb62fe27833c589ab8a900c202fb0e019de76ab7e50b6d87098a3cebc2f79c04f0737cf4551bc209457599e0806b6e36ce9725bddf06a7aeb0fd5ed7127a02c48981c90ed1e47f9492d32a73c930eea823fb8ce7a23f73cff3381890c01c94797dfa4fbc9b1b226d65972c80acf1d10a471801c48aeccbd7b325b50b408b20977c6445febc6a1c8a849bdbc64b4497f08b141418164b31bc999b9869f01f138a142aa38e4454d217e8532290f20e211c5578639f0f7a8a381e02f667020bffc27977f2ca0a735ccf8d946b981443d1026a8b70be166b50005e0357789067cd8f68081a941e1068961830f36a73e24cbc2cd1dac1cb4a781475b5a95c3"
-        });
-
-        uint256 validationData = mockModule.validateSignaturesExternal(userOp);
-        address authorizer = address(uint160(validationData));
-        uint48 validUntil = uint48(validationData >> 160);
-        uint48 validAfter = uint48(validationData >> 208);
-
-        console.log("authorizer: %s", authorizer);
-        console.log("validUntil: %s", validUntil);
-        console.log("validAfter: %s", validAfter);
-    }
-
-    function test_decodeAggregatedSignature() public {
-        PBHPayload[] memory pbhPayloads = abi.decode(opsPerAggregator[i].signature, (PBHPayload[]));
-    }
-
     function testFuzz_ValidateUserOpSignature(
         uint256 root,
         uint256 pbhExternalNullifier,
