+++ world-chain-builder/crates/world/test-utils/src/utils.rs
@@ -4,6 +4,7 @@ use alloy_network::TxSigner;
 use alloy_primitives::{
     aliases::U48, bytes, fixed_bytes, keccak256, Address, Bytes, ChainId, FixedBytes, TxKind, U256,
 };
+use alloy_primitives::{B256, U128, U64, U8};
 use alloy_rlp::Encodable;
 use alloy_signer::SignerSync;
 use alloy_signer_local::{coins_bip39::English, PrivateKeySigner};
@@ -16,6 +17,7 @@ use reth_primitives::transaction::SignedTransactionIntoRecoveredExt;
 use semaphore::identity::Identity;
 use semaphore::poseidon_tree::LazyPoseidonTree;
 use semaphore::{hash_to_field, Field};
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
 use std::{str::FromStr, sync::LazyLock};
 use world_chain_builder_pbh::external_nullifier::ExternalNullifier;
@@ -152,7 +154,7 @@ pub fn user_op(
     calldata: Bytes,
     #[builder(default = fixed_bytes!("000000000000000000000000000fffd30000000000000000000000000000C350"))]
     account_gas_limits: FixedBytes<32>,
-    #[builder(default = U256::from(60232))] pre_verification_gas: U256,
+    #[builder(default = U256::from(500836))] pre_verification_gas: U256,
     #[builder(default = fixed_bytes!("0000000000000000000000003B9ACA0000000000000000000000000073140B60"))]
     gas_fees: FixedBytes<32>,
     #[builder(default = Bytes::default())] paymaster_and_data: Bytes,
@@ -213,7 +215,6 @@ pub fn user_op(
 
     user_op.signature = Bytes::from(uo_sig);
 
-    println!("UserOp: {:?}", user_op);
     (user_op, payload)
 }
 
@@ -377,9 +378,9 @@ impl From<PbhPayload> for PBHPayload {
     }
 }
 
-impl Into<alloy_rpc_types::PackedUserOperation> for PackedUserOperation {
-    fn into(self) -> alloy_rpc_types::PackedUserOperation {
-        alloy_rpc_types::PackedUserOperation {
+impl Into<RpcUserOperationV0_7> for PackedUserOperation {
+    fn into(self) -> RpcUserOperationV0_7 {
+        RpcUserOperationV0_7 {
             sender: self.sender,
             nonce: self.nonce,
             factory: None,
@@ -402,6 +403,8 @@ impl Into<alloy_rpc_types::PackedUserOperation> for PackedUserOperation {
             paymaster_data: None,
             paymaster_post_op_gas_limit: None,
             paymaster_verification_gas_limit: None,
+            aggregator: Some(PBH_DEV_SIGNATURE_AGGREGATOR),
+            eip7702_auth: None,
         }
     }
 }
@@ -410,6 +413,120 @@ pub fn hash_user_op(user_op: &PackedUserOperation) -> Field {
     let hash = SolValue::abi_encode_packed(&(&user_op.sender, &user_op.nonce, &user_op.callData));
 
     hash_to_field(hash.as_slice())
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, Eq, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct RpcUserOperationByHash {
+    /// The full user operation
+    pub user_operation: RpcUserOperation,
+    /// The entry point address this operation was sent to
+    pub entry_point: RpcAddress,
+    /// The number of the block this operation was included in
+    pub block_number: Option<U256>,
+    /// The hash of the block this operation was included in
+    pub block_hash: Option<B256>,
+    /// The hash of the transaction this operation was included in
+    pub transaction_hash: Option<B256>,
+}
+
+/// authorization tuple for 7702 txn support
+#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct RpcEip7702Auth {
+    /// The chain ID of the authorization.
+    pub chain_id: U64,
+    /// The address of the authorization.
+    pub address: Address,
+    /// The nonce for the authorization.
+    pub nonce: U64,
+    /// signed authorizzation tuple.
+    pub y_parity: U8,
+    /// signed authorizzation tuple.
+    pub r: U256,
+    /// signed authorizzation tuple.
+    pub s: U256,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct RpcAddress(Address);
+
+impl Serialize for RpcAddress {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        serializer.serialize_str(&self.0.to_checksum(None))
+    }
+}
+
+impl<'de> Deserialize<'de> for RpcAddress {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let address = Address::deserialize(deserializer)?;
+        Ok(RpcAddress(address))
+    }
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize, Eq, PartialEq)]
+#[serde(untagged)]
+pub enum RpcUserOperation {
+    V0_6(RpcUserOperationV0_6),
+    V0_7(RpcUserOperationV0_7),
+}
+
+/// User operation definition for RPC
+#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct RpcUserOperationV0_6 {
+    sender: RpcAddress,
+    nonce: U256,
+    init_code: Bytes,
+    call_data: Bytes,
+    call_gas_limit: U128,
+    verification_gas_limit: U128,
+    pre_verification_gas: U128,
+    max_fee_per_gas: U128,
+    max_priority_fee_per_gas: U128,
+    paymaster_and_data: Bytes,
+    signature: Bytes,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    eip7702_auth: Option<RpcEip7702Auth>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    aggregator: Option<Address>,
+}
+
+/// User operation definition for RPC inputs
+#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct RpcUserOperationV0_7 {
+    sender: Address,
+    nonce: U256,
+    call_data: Bytes,
+    call_gas_limit: U128,
+    verification_gas_limit: U128,
+    pre_verification_gas: U256,
+    max_priority_fee_per_gas: U128,
+    max_fee_per_gas: U128,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    factory: Option<Address>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    factory_data: Option<Bytes>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    paymaster: Option<Address>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    paymaster_verification_gas_limit: Option<U128>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    paymaster_post_op_gas_limit: Option<U128>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    paymaster_data: Option<Bytes>,
+    signature: Bytes,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    eip7702_auth: Option<RpcEip7702Auth>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    aggregator: Option<Address>,
 }
 
 #[cfg(test)]
